requirements.txt
Flask==3.1.0
flask-cors==5.0.1
numpy==2.2.4
opencv-python==4.11.0.86
pillow==11.2.1
